Creating a usable API (#1713)
* Initial stab at API structuring

* Throwing all the things into the API*
Eliminated all internal imports
Also added some helpful comments
*except for the ritual stuff

* Reducing the API
Threw back the altar/incense/unnecessary items to main
Added in a functional API instance

* API cleanup
Removing all the unnecessities
Smushed and vaporized some redundant recipe stuffs

* Made API dummy instances
Refactor packaging

+++ src/main/java/wayoftime/bloodmagic/client/hud/Elements.java
@@ -31,7 +31,7 @@ public class Elements
 				// Current tier
 				information.accept(Pair.of(new Sprite(new ResourceLocation(BloodMagic.MODID, "textures/gui/widgets.png"), 0, 46, 16, 16), altar -> altar == null
 						? "IV"
-						: NumeralHelper.toRoman(altar.getTier().toInt())));
+						: NumeralHelper.toRoman(altar.getTier())));
 				// Stored/Capacity
 				information.accept(Pair.of(new Sprite(new ResourceLocation(BloodMagic.MODID, "textures/gui/widgets.png"), 16, 46, 16, 16), altar -> String.format("%d/%d", altar == null
 						? 0
@@ -47,7 +47,7 @@ public class Elements
 				// Current tier
 				information.accept(Pair.of(new Sprite(new ResourceLocation(BloodMagic.MODID, "textures/gui/widgets.png"), 0, 46, 16, 16), altar -> altar == null
 						? "IV"
-						: NumeralHelper.toRoman(altar.getTier().toInt())));
+						: NumeralHelper.toRoman(altar.getTier())));
 				// Stored/Capacity
 				information.accept(Pair.of(new Sprite(new ResourceLocation(BloodMagic.MODID, "textures/gui/widgets.png"), 16, 46, 16, 16), altar -> String.format("%d/%d", altar == null
 						? 0
